+++ src/Converter/Time/DegradedTimeConverter.php
@@ -14,47 +14,10 @@ declare(strict_types=1);
 
 namespace Ramsey\Uuid\Converter\Time;
 
-use Ramsey\Uuid\Converter\TimeConverterInterface;
-use Ramsey\Uuid\Exception\UnsatisfiedDependencyException;
-
 /**
- * DegradedTimeConverter is chosen if all other options for large integer
- * support are unavailable. This exists to throw exceptions if these methods
- * are called on systems that do not have support for large integers.
+ * @deprecated DegradedTimeConverter is no longer necessary for converting
+ *     time on 32-bit systems. Transition to {@see GenericTimeConverter}.
  */
-class DegradedTimeConverter implements TimeConverterInterface
+class DegradedTimeConverter extends BigNumberTimeConverter
 {
-    /**
-     * @throws UnsatisfiedDependencyException if the chosen converter is not present
-     *
-     * @inheritDoc
-     *
-     * @psalm-pure
-     */
-    public function calculateTime(string $seconds, string $microSeconds): array
-    {
-        throw new UnsatisfiedDependencyException(
-            'Cannot call calculateTime using the DegradedTimeConverter; '
-            . 'please choose a converter with support for large integers; '
-            . 'refer to the ramsey/uuid wiki for more information: '
-            . 'https://github.com/ramsey/uuid/wiki'
-        );
-    }
-
-    /**
-     * @throws UnsatisfiedDependencyException if the chosen converter is not present
-     *
-     * @inheritDoc
-     *
-     * @psalm-pure
-     */
-    public function convertTime(string $timestamp): string
-    {
-        throw new UnsatisfiedDependencyException(
-            'Cannot call convertTime using the DegradedTimeConverter; '
-            . 'please choose a converter with support for large integers; '
-            . 'refer to the ramsey/uuid wiki for more information: '
-            . 'https://github.com/ramsey/uuid/wiki'
-        );
-    }
 }