+++ src/FeatureSet.php
@@ -15,16 +15,13 @@ declare(strict_types=1);
 namespace Ramsey\Uuid;
 
 use Ramsey\Uuid\Builder\DefaultUuidBuilder;
-use Ramsey\Uuid\Builder\DegradedUuidBuilder as Rfc4122DegradedUuidBuilder;
 use Ramsey\Uuid\Builder\FallbackBuilder;
 use Ramsey\Uuid\Builder\UuidBuilderInterface;
 use Ramsey\Uuid\Codec\CodecInterface;
 use Ramsey\Uuid\Codec\GuidStringCodec;
 use Ramsey\Uuid\Codec\StringCodec;
-use Ramsey\Uuid\Converter\Number\DegradedNumberConverter;
 use Ramsey\Uuid\Converter\Number\GenericNumberConverter;
 use Ramsey\Uuid\Converter\NumberConverterInterface;
-use Ramsey\Uuid\Converter\Time\DegradedTimeConverter;
 use Ramsey\Uuid\Converter\Time\GenericTimeConverter;
 use Ramsey\Uuid\Converter\Time\PhpTimeConverter;
 use Ramsey\Uuid\Converter\TimeConverterInterface;
@@ -33,9 +30,7 @@ use Ramsey\Uuid\Generator\RandomGeneratorFactory;
 use Ramsey\Uuid\Generator\RandomGeneratorInterface;
 use Ramsey\Uuid\Generator\TimeGeneratorFactory;
 use Ramsey\Uuid\Generator\TimeGeneratorInterface;
-use Ramsey\Uuid\Guid\DegradedGuidBuilder;
 use Ramsey\Uuid\Guid\GuidBuilder;
-use Ramsey\Uuid\Nonstandard\DegradedUuidBuilder as NonstandardDegradedUuidBuilder;
 use Ramsey\Uuid\Math\BrickMathCalculator;
 use Ramsey\Uuid\Math\CalculatorInterface;
 use Ramsey\Uuid\Nonstandard\UuidBuilder as NonstandardUuidBuilder;
@@ -328,33 +323,14 @@ class FeatureSet
      */
     private function buildUuidBuilder(bool $useGuids = false): UuidBuilderInterface
     {
-        if ($this->is64BitSystem() && $useGuids) {
+        if ($useGuids) {
             return new GuidBuilder($this->numberConverter, $this->timeConverter);
         }
 
-        if ($this->is64BitSystem()) {
-            return new FallbackBuilder([
-                new DefaultUuidBuilder($this->numberConverter, $this->timeConverter),
-                new NonstandardUuidBuilder($this->numberConverter, $this->timeConverter),
-            ]);
-        }
-
-        if ($useGuids) {
-            return new DegradedGuidBuilder($this->numberConverter, $this->timeConverter);
-        }
-
         return new FallbackBuilder([
-            new Rfc4122DegradedUuidBuilder($this->numberConverter, $this->timeConverter),
-            new NonstandardDegradedUuidBuilder($this->numberConverter, $this->timeConverter),
+            new DefaultUuidBuilder($this->numberConverter, $this->timeConverter),
+            new NonstandardUuidBuilder($this->numberConverter, $this->timeConverter),
         ]);
-    }
-
-    /**
-     * Returns true if moontoast/math is available
-     */
-    private function hasBigNumber(): bool
-    {
-        return class_exists('Moontoast\Math\BigNumber') && !$this->disableBigNumber;
     }
 
     /**